Turn into class Grammar
Makes for a really ugly diff :/ I swear I didn't change the logic.
Ideally I'd have tests to prove my refactoring correct, but this is
precisely a step towards adding tests (and anyway the code is not
mature, so it's not a problem to change the set of bugs yet).